easyparking-ml/requirements.txt
minio==7.1.12
rabbitmq==0.2.0
autopep8==2.0.0
pika==1.3.1
absl-py==1.3.0
alabaster==0.7.12
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
attrs==22.1.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
cycler==0.11.0
Cython==0.29.32
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
docutils==0.19
entrypoints==0.4
fastjsonschema==2.16.2
fonttools==4.38.0
gast==0.5.3
grpcio==1.50.0
h5py==2.10.0
idna==3.4
imageio==2.22.4
imagesize==1.4.1
imgaug==0.4.0
importlib-metadata==5.0.0
importlib-resources==5.10.0
ipykernel==6.16.2
ipyparallel==8.4.1
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.17.0
jupyter-client==7.4.4
jupyter-core==4.11.2
jupyter-server==1.23.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
Keras==2.0.8
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==2.0.4
mock==4.0.3
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.4
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.6.3
nose==1.3.7
notebook==6.5.2
notebook-shim==0.2.2
numpy==1.21.6
opencv-python==4.6.0.66
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
pkgutil-resolve-name==1.3.10
prometheus-client==0.15.0
prompt-toolkit==3.0.32
protobuf==3.20.0
psutil==5.9.4
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.6
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
requests==2.28.1
scikit-image==0.19.3
scipy==1.7.3
Send2Trash==1.8.0
Shapely==1.8.5.post1
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
sphinx==5.3.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==2.1.0
terminado==0.17.0
testpath==0.6.0
tifffile==2021.11.2
tinycss2==1.2.1
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
typing-extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
widgetsnbextension==4.0.3
zipp==3.10.0
